feat: refactor JobList component to filter jobs by tags and enhance utility functions
